Simplify logs viewer pattern for publication
