+++ examples/recipes/backend/docker-compose.yaml
@@ -59,3 +59,5 @@ services:
       - ${DEMO_API_PORT:-3123}:3123
     depends_on:
       - postgres
+    volumes:
+      - ./demo-server:/usr/src/app
